apps/web/requirements.txt
bcrypt==4.1.3
fastapi==0.115.5
python-jose==3.3.0
redis==5.0.7
tortoise-orm==0.21.3
concurrent-log-handler==0.9.25
uvicorn[standard]~=0.30.1
starlette~=0.40.0
oss2~=2.18.6
PyYAML~=6.0.1
aiohttp~=3.10.2
typing_inspect~=0.9.0
dependency-injector~=4.43.0
aiomysql~=0.2.0
elasticsearch~=8.15.1
aiokafka~=0.11.0
typing-inspect~=0.9.0
typing_extensions~=4.12.2
aiosmtplib~=3.0.2
alibabacloud_dysmsapi20170525==3.0.0
kafka-python~=2.0.2
six~=1.16.0
pytz~=2024.1
alibabacloud_tea_openapi~=0.3.12
alibabacloud_tea_util~=0.3.13
pydantic~=2.9.2
gunicorn~=23.0.0
python-multipart~=0.0.9